+++ src/main/kotlin/team/mozu/dsm/application/port/in/team/TeamParticipationUseCase.kt
@@ -1,4 +1,4 @@
-package team.mozu.dsm.application.team.port.`in`
+package team.mozu.dsm.application.port.`in`.team
 
 import team.mozu.dsm.adapter.`in`.team.dto.request.TeamParticipationRequest
 import team.mozu.dsm.adapter.`in`.team.dto.response.TeamTokenResponse
